[refactor] 🔧 Fix typing refactor brought-in issues
