+++ docs/tutorial/convert.md
@@ -2,7 +2,6 @@
 
 The module focuses on automatically converting research models to serialized and optimized models from Python code.
 
-
 We currently support 2 research frameworks:
 
 -   TensorFlow
@@ -16,10 +15,13 @@ And MLModelCI supports the following conversions:
 -   Tensorflow -> TensorRT format
 -   ONNX -> TensorRT format
 
-Upon receiving your model registration, MLModelCI will help you to convert your model to as many as possible optimized formats. You can also use the following APIs to convert models manually.
+Upon receiving your model registration, MLModelCI will help you to convert your model to as many as possible optimized 
+formats. You can also use the following APIs to convert models manually.
 
-
-These APIs will save the converted model in the given `saved_path` and return success or failure status of the conversion. We restrict to use the standard `saved_path` generated by `modelci.hub.utils.generate_path(...)` or path according to the rules (See [Tricks with Model Saved Path](./register.md#tricks-with-model-saved-path)). This path format can make your life easier.
+These APIs will save the converted model in the given `saved_path` and return success or failure status of the 
+conversion. We restrict to use the standard `saved_path` generated by `modelci.hub.utils.generate_path(...)` or path 
+according to the rules (See [Tricks with Model Saved Path](./register.md#tricks-with-model-saved-path)). This path 
+format can make your life easier.
 
 ### 1. TorchScript Conversion
 
@@ -58,13 +60,15 @@ ONNXConverter.from_torch_module(torch_module, saved_path, input_shape, batch_siz
 
 ### 4. TRT Conversion
 
-Different from the above converters, TRT accepts both tfsavedmodel and onnx formats for further optimization. In practice, if you use this pipeline: PyTorch -> onnx -> trt, you can get the best performance. However, in most cases, PyTorch model can not be further converted.
+Different from the above converters, TRT accepts both TensorFlow SavedModel and ONNX formats for further optimization. 
+In practice, if you use this pipeline: PyTorch -> ONNX -> TRT, you can get the best performance. However, in most 
+cases, PyTorch model can not be further converted.
 
 #### From TF Savedmodel to TF-TRT
 
 ```python
 from modelci.hub.converter import TRTConverter
-from modelci.persistence.bo import IOShape
+from modelci.types.bo import IOShape
 
 tf_path = ...
 trt_path = ...
@@ -78,7 +82,7 @@ TRTConverter.from_saved_model(tf_path, trt_path, inputs=inputs, outputs=outputs)
 
 ```python
 from modelci.hub.converter import TRTConverter
-from modelci.persistence.bo import IOShape
+from modelci.types.bo import IOShape
 
 onnx_path = ...
 save_path = ...
