+++ modelci/metrics/cadvisor/CLIENT.md
@@ -26,38 +26,58 @@ You can use `-h` or `--help` to get more information.
 
 Init the cAdvisor client, you can set the API informations here.
 
+With default settings:
 ```python
-cAdvisor = CAdvisor() # init with default settings
-cAdvisor = CAdvisor(base_link='http://localhost:8080/api/', ersion='v1.2', query_object='docker') 
-# init with specific API dettings
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor()
+```
+
+With specific API settings:
+```python
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(base_link='http://localhost:8080/api/', ersion='v1.2', query_object='docker') 
 ```
 
 #### Request the Host Machine Information
 
 You can also get the host machine info by:
-
 ```python
-machine_info = cAdvisor.machine_info()
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...) 
+
+machine_info = c_advisor.machine_info()
 ```
 
 #### Get Basic Information from Running Containers
 
 If you just want to check the information of running containers, you can do:
-
 ```python
-data_running_containers = cAdvisor.get_running_container()
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...) 
+
+data_running_containers = c_advisor.get_running_container()
 # it also receives an input screened dict
-all_information_by_id = cAdvisor.request_by_id("afcb380e62bbec4fb7992cda3c986d387b5bc137fdea7dc0d1d13448012d5a5d")
-data_running_container_by_id = cAdvisor.get_running_container(all_information_by_id)
+all_information_by_id = c_advisor.request_by_id('afcb380e62bbec4fb7992cda3c986d387b5bc137fdea7dc0d1d13448012d5a5d')
+data_running_container_by_id = c_advisor.get_running_container(all_information_by_id)
 ```
 
 #### Request All Container Data
 
 Request the JSON data, each request will return the information in the last one minute (60 seconds), if the container just started within 1 min, the request will return all of them.
-
 ```python
-all_data = cAdvisor.request_all() # request all the data at once
-all_stat = cAdvisor.get_stats() # request all the stats per second for all running containers
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...) 
+
+# request all the data at once
+all_data = c_advisor.request_all()
+
+# request all the stats per second for all running containers
+all_stat = c_advisor.get_stats()
 ```
 
 #### Request Data from Specific Containers
@@ -65,19 +85,28 @@ all_stat = cAdvisor.get_stats() # request all the stats per second for all runni
 You can set filters to the request, if we only need information from containers with image `ubuntu:16.04`:
 
 ```python
-all_information_ubuntu = cAdvisor.request_by_image("ubuntu:16.04")
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...)
+
+all_information_ubuntu = c_advisor.request_by_image("ubuntu:16.04")
 ```
 
 We can also request data by other fields like name(id or image name), id, or just all the information expect cadvisor itself.
 
 ```python
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...)
+
 # request by id
-all_information_by_id = cAdvisor.request_by_id("afcb380e62bbec4fb7992cda3c986d387b5bc137fdea7dc0d1d13448012d5a5d") 
+all_information_by_id = c_advisor.request_by_id('afcb380e62bbec4fb7992cda3c986d387b5bc137fdea7dc0d1d13448012d5a5d') 
 # request by name or id
-all_data_cadvisor = cAdvisor.request_by_name("cadvisor") 
-all_data_cadvisor = cAdvisor.request_by_name("afcb380e62bbec4fb7992cda3c986d387b5bc137fdea7dc0d1d13448012d5a5d") 
-# request all the container informations without cAdvisor
-data_all_without_cadvisor = cAdvisor.request_without_cadvisor()
+all_data_cadvisor = c_advisor.request_by_name('cadvisor') 
+all_data_cadvisor = c_advisor.request_by_name('afcb380e62bbec4fb7992cda3c986d387b5bc137fdea7dc0d1d13448012d5a5d') 
+
+# request all the container information without cAdvisor
+data_all_without_cadvisor = c_advisor.request_without_cadvisor()
 ```
 
 #### Request Data by Specific Metric
@@ -85,15 +114,23 @@ data_all_without_cadvisor = cAdvisor.request_without_cadvisor()
 If you want to get some specific data by metrics like: diskio, cpu, etc.:
 
 ```python
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...)
+
 # The field can be: diskio, cpu, diskio, memory, network, filesystem, task_stats, processes, accelerators (Nvidia GPU only)
-data_diskio_all = cAdvisor.get_specific_metrics(metric='diskio')
+data_diskio_all = c_advisor.get_specific_metrics(metric='diskio')
 ```
 
 It also receives a screened dict if you want the specific metric from some specific containers:
 
 ```python
-all_information_ubuntu = cAdvisor.request_by_image("ubuntu:16.04")
-data_cpu_ubuntu = cAdvisor.get_specific_metrics(input_dict=all_information_ubuntu, metric='cpu')
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...)
+
+all_information_ubuntu = c_advisor.request_by_image('ubuntu:16.04')
+data_cpu_ubuntu = c_advisor.get_specific_metrics(input_dict=all_information_ubuntu, metric='cpu')
 ```
 
 #### Export the Data 
@@ -102,7 +139,12 @@ All the data here are Python dicts, you can use json to dump them out if you wan
 
 ```python
 import json
-all_data = cAdvisor.request_all()
+
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...)
+
+all_data = c_advisor.request_all()
 json.dumps(all_data)
 ```
 
@@ -132,9 +174,14 @@ Currently, we have three containers based on the GPU accelerators:
 
 We use below code to get data
 
-```python 
-cAdvisor = CAdvisor()
-example_cuda_101base = cAdvisor.request_by_image("nvidia/cuda:10.1-base")
+```python
+import json
+
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
+
+c_advisor = CAdvisor(...)
+
+example_cuda_101base = c_advisor.request_by_image('nvidia/cuda:10.1-base')
 with open('./example_data/example_cuda_101base.json', 'w') as f:
     json.dump(example_cuda_101base, f)
 ```
