+++ modelci/metrics/cadvisor/README.md
@@ -24,7 +24,7 @@ Using following code, you can get the required formated data from [cAdvisor](htt
 
 ```python
  
-from cadvisor import CAdvisor
+from modelci.metrics.cadvisor.cadvisor import CAdvisor
 
 client = CAdvisor(base_link='') # link to your remote 
 tensorflow_resnet_latest = client.request_by_image('bitnami/tensorflow-serving:latest') # set filters
